+++ Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 authors = ["andreespirela <andreespirela@outlook.com>", "Divy Srivastava <dj.srivastava23@gmail.com>"]
 edition = "2021"
 resolver = "2"
+build = "build.rs"
 
 [dependencies]
 v8 = "0.34.0"
@@ -18,6 +19,11 @@ async-trait = "0.1.51"
 pico-args = "0.4.2"
 colored = "2"
 
-[[bin]]
-name = "snapshot"
-path = "src/bin/snapshot.rs"
+[build-dependencies]
+v8 = "0.34.0"
+deno_core = "0.107.0"
+deno_crypto = "0.39.0"
+deno_url = "0.25.0"
+deno_web = "0.56.0"
+deno_webidl = "0.25.0"
+serde = { version = "1", features = ["derive"] }
