--- src/lib.rs
@@ -1,2 +1,0 @@
-pub mod runtime;
-mod snapshot_js;
